pyproject.toml
[project]
name = "drupal-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.1.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
]
